% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_conversions.R
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Hex to RGB}
\usage{
hex_to_rgb(hex)
}
\arguments{
\item{hex}{vector of hex colours}
}
\description{
converts hex string colours to RGB
}
\examples{
colour_to_rgb(c("#00FF00","#FF00FF"))

}
